+++ src/features/api/schemas/user.schemas.ts
@@ -5,6 +5,8 @@ import {
 	IUserLoginPayload,
 	IUserRefreshPayload,
 	IUserTokenResponse,
+	IUserAPIKey,
+	IUserAPIKeyCreatePayload,
 } from "@/features/api/userData.types";
 
 export const LoginPayloadSchema: z.ZodType<IUserLoginPayload> = z.object({
@@ -74,6 +76,21 @@ export const UserVerifyEmailResponseSchema = z.object({
 	message: z.string(),
 });
 
+const UserAPIKeySchema: z.ZodType<IUserAPIKey> = z.object({
+	name: z.string(),
+	key: z.string(),
+	created_date: z.coerce.date(),
+	last_activity: z.coerce.date().nullable(),
+});
+
+export const UserAPIKeyPayloadSchema: z.ZodType<IUserAPIKey[]> =
+	z.array(UserAPIKeySchema);
+
+export const UserAPIKeyCreatePayloadSchema: z.ZodType<IUserAPIKeyCreatePayload> =
+	z.object({
+		keyname: z.string().min(1).max(100),
+	});
+
 export type LoginPayloadType = z.infer<typeof LoginPayloadSchema>;
 export type TokenResponseType = z.infer<typeof TokenResponseSchema>;
 export type RefreshPayloadType = z.infer<typeof RefreshPayloadSchema>;
@@ -92,4 +109,8 @@ export type UserVerifyEmailPayloadType = z.infer<
 >;
 export type UserVerifyEmailResponseType = z.infer<
 	typeof UserVerifyEmailResponseSchema
+>;
+export type UserAPIKeyPayloadType = z.infer<typeof UserAPIKeyPayloadSchema>;
+export type UserAPIKeyCreatePayloadType = z.infer<
+	typeof UserAPIKeyCreatePayloadSchema
 >;